+++ packages/admin/admin-color-picker/src/index.ts
@@ -1,2 +1,11 @@
-export { ColorPickerProps, default as FinalFormColorPicker } from "./core/ColorPicker";
-export { ColorPickerClassKey } from "./core/ColorPicker.styles";
+export { ColorPicker, ColorPickerProps } from "./ColorPicker";
+export { ColorPickerClassKey } from "./ColorPicker.styles";
+export {
+    ColorPickerPreview,
+    ColorPickerPreviewColorProps,
+    ColorPickerPreviewEmptyProps,
+    ColorPickerPreviewInvalidProps,
+    ColorPickerPreviewProps,
+} from "./ColorPickerPreview";
+export { ColorPickerPreviewClassKey } from "./ColorPickerPreview.styles";
+export { FinalFormColorPicker, FinalFormColorPickerProps } from "./FinalFormColorPicker";